moving the main funcitons to hw layer

+++ firmware/dev/f4dev/cubemx/Src/usbd_cdc_if.c
@@ -22,8 +22,8 @@
 #include "usbd_cdc_if.h"
 
 /* USER CODE BEGIN INCLUDE */
-#include "hw_usbMsgQueue.h"
-
+#include "hw_usb.h"
+#include "cmsis_os2.h"
 /* USER CODE END INCLUDE */
 
 /* Private typedef -----------------------------------------------------------*/
@@ -326,7 +326,7 @@ static int8_t CDC_TransmitCplt_FS(uint8_t *Buf, uint32_t *Len, uint8_t epnum)
     UNUSED(Len);
     UNUSED(epnum);
     /* USER CODE END 13 */
-    return result;
+    return (int8_t)result;
 }
 
 /* USER CODE BEGIN PRIVATE_FUNCTIONS_IMPLEMENTATION */
